oracle-ai-data-platform/setup/requirements.txt
oci-cli
python-dotenv==1.1.0
oracle-ads==2.13.8
oracledb==3.1.0
pandas==2.2.3
PyMuPDF==1.25.5

langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34

pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2

tensorflow==2.18.0
transformers==4.48.3

graphviz==0.20.3
htbuilder==0.8.0

streamlit==1.44.1
st-annotated-text==4.0.2
streamlit-float==0.3.5

presidio-analyzer==2.2.358
presidio-anonymizer==2.2.358


oci-ai-speech-realtime
PyAudio>=0.2.14
websockets>=11.0.3